+++ errors/types_generated.go
@@ -13,8 +13,10 @@ var (
 	Saml2                       = Definitions[8]
 	Saml                        = Definitions[9]
 	BetaInsufficientPermissions = Definitions[10]
-	Sequence                    = Definitions[12]
-	UnexpectedStatus            = Definitions[13]
+	BackendOpenapiValidation    = Definitions[12]
+	BackendTimeout              = Definitions[13]
+	Sequence                    = Definitions[14]
+	UnexpectedStatus            = Definitions[15]
 )
 
 // typeDefinitions holds all related error definitions which are
@@ -35,7 +37,9 @@ var types = typeDefinitions{
 	"saml2":                          Saml2,
 	"saml":                           Saml,
 	"beta_insufficient_permissions":  BetaInsufficientPermissions,
-	"backend_validation":             BackendValidation,
+	"backend":                        Backend,
+	"backend_openapi_validation":     BackendOpenapiValidation,
+	"backend_timeout":                BackendTimeout,
 	"sequence":                       Sequence,
 	"unexpected_status":              UnexpectedStatus,
 }
@@ -46,3 +50,7 @@ func IsKnown(errorType string) bool {
 	_, known := types[errorType]
 	return known
 }
+
+// SuperTypesMapsByContext holds maps for error super-types to sub-types
+// by a given context block type (e.g. api or endpoint).
+var SuperTypesMapsByContext = map[string]map[string][]string{"api": map[string][]string{"*": []string{"beta_insufficient_permissions", "backend_openapi_validation", "backend_timeout"}, "access_control": []string{"beta_insufficient_permissions"}, "backend": []string{"backend_openapi_validation", "backend_timeout"}}, "endpoint": map[string][]string{"*": []string{"beta_insufficient_permissions", "backend_openapi_validation", "backend_timeout", "sequence", "unexpected_status"}, "access_control": []string{"beta_insufficient_permissions"}, "backend": []string{"backend_openapi_validation", "backend_timeout"}, "endpoint": []string{"sequence", "unexpected_status"}}}